+++ CHANGELOG.md
@@ -6,6 +6,9 @@ All notable changes to this project will be documented in this file.
 
 [Full diff](https://github.com/sider/runners/compare/0.30.0...HEAD)
 
+- **Brakeman** Syntax highlight for warning message [#1309](https://github.com/sider/runners/pull/1309)
+- **stylelint** Add `stylelint-order` to pre-installed list [#1310](https://github.com/sider/runners/pull/1310)
+
 ## 0.30.0
 
 [Full diff](https://github.com/sider/runners/compare/0.29.3...0.30.0)